requirements.txt
# VC Music Bot — Requirements
# Install: pip install -r requirements.txt

pyrogram==2.0.106
TgCrypto
pytgcalls>=0.9.0
yt-dlp>=2024.1.1
aiosqlite>=0.19.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
